Implement embedding picture title into image file
The --embed-meta parameter will create a copy of the image
with picture title embedded on top right corner of the file.
The title size is adjusted to the current screen scaling factor.
The image is also resized to match the screen resolution to make
sure the title text will not get resized or cropped by Windows.